pyproject.toml
[project]
name = "tli-self-service-ms-teams"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.1",
    "httpx>=0.28.1",
    "msal>=1.34.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
